Add support for Vagrant file output
In addition to creating the WDS image toolchain, also output a .box file
that can be used for Vagrant.

To use, simply download the .box and run `vagrant box add <file>` and
then you can run `vagrant up` to have a fully functional testing
machine.

Support will be added in the future for integrating Tech Tools via the
Vagrantfile; Orca, RSAT, etc can be piped in during the first run of the
virtual machine.