Fix minor bug in Windows Update script
The Count and other properties of the items were not accessible if there
is only 1 item in the array. This fix ensures that all properties can be
read.

This was tested and discovered in earlier Windows 8.1 builds. This
results in a longer update time, which proves that there are now
more updates being installed.